refactor(menu): rename menu.type to types
