+++ packages/components/menu/src/sub-menu.vue
@@ -24,7 +24,7 @@ import type {
   ISubMenuProps,
   MenuProvider,
   SubMenuProvider,
-} from './menu.type'
+} from './types'
 
 export default defineComponent({
   name: 'ElSubMenu',
