+++ packages/components/menu/src/use-menu-color.ts
@@ -1,7 +1,7 @@
 import { computed } from 'vue'
 import { darken } from '@element-plus/utils/color'
 
-import type { IMenuProps } from './menu.type'
+import type { IMenuProps } from './types'
 
 export default function useMenuColor(props: IMenuProps) {
   const menuBarColor = computed(() => {
